Refactor BrowsePhoto into PhotoAnalizer
